requirements.txt
# Core packages
streamlit==1.41.1
pandas==2.2.3
numpy==2.0.2
pillow==11.1.0

# TensorFlow stack
tensorflow==2.18.0
tensorflow-intel==2.18.0          # optional untuk Windows Intel CPU
keras==3.8.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0

# Dependencies TensorFlow
absl-py==2.1.0
astunparse==1.6.3
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
libclang==18.1.1
ml-dtypes==0.4.1
namex==0.0.8
opt-einsum==3.4.0
optree==0.14.0
protobuf==5.29.3
six==1.17.0
termcolor==2.5.0
typing_extensions==4.12.2
wrapt==1.17.2

# Streamlit dependencies
altair==5.5.0
blinker==1.9.0
gitpython==3.1.44
pydeck==0.9.1
watchdog==6.0.0
requests==2.32.3
tornado==6.4.2
toml==0.10.2
jinja2==3.1.5
markdown==3.7
rich==13.9.4
werkzeug==3.1.3

# Utilities
packaging==24.2
cachetools==5.5.1
python-dateutil==2.9.0.post0
tzdata==2025.1
urllib3==2.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
